+++ unique_toolkit/unique_toolkit/tools/tool_manager.py
@@ -4,6 +4,7 @@ from pydantic import BaseModel, Field
 from unique_toolkit.app.schemas import ChatEvent
 from unique_toolkit.language_model.schemas import (
     LanguageModelFunction,
+    LanguageModelTool,
     LanguageModelToolDescription,
 )
 from unique_toolkit.tools.config import ToolBuildConfig
@@ -12,6 +13,7 @@ from unique_toolkit.tools.schemas import ToolCallResponse, ToolPrompts
 from unique_toolkit.tools.tool import Tool
 from unique_toolkit.tools.tool_progress_reporter import ToolProgressReporter
 from unique_toolkit.tools.utils.execution.execution import Result, SafeTaskExecutor
+from unique_toolkit.evals.schemas import EvaluationMetricName
 
 
 class ForcedToolOption:
@@ -67,16 +69,18 @@ class ToolManager:
         tool_progress_reporter: ToolProgressReporter,
     ):
         self._logger = logger
-        self._config  = config
+        self._config = config
         self._tool_progress_reporter = tool_progress_reporter
         self._tools = []
         self._tool_choices = event.payload.tool_choices
         self._disabled_tools = event.payload.disabled_tools
+        # this needs to be a set of strings to avoid duplicates
+        self._tool_evaluation_check_list: set[EvaluationMetricName] = set()
         self._init__tools(event)
 
-    def _init__tools(self,  event: ChatEvent) -> None:
+    def _init__tools(self, event: ChatEvent) -> None:
         tool_choices = self._tool_choices
-        tool_configs = self._config .tools
+        tool_configs = self._config.tools
         self._logger.info("Initializing tool definitions...")
         self._logger.info(f"Tool choices: {tool_choices}")
         self._logger.info(f"Tool configs: {tool_configs}")
@@ -105,6 +109,9 @@ class ToolManager:
 
             self._tools.append(t)
 
+    def get_evaluation_check_list(self) -> list[EvaluationMetricName]:
+        return list(self._tool_evaluation_check_list)
+
     def log_loaded_tools(self):
         self._logger.info(f"Loaded tools: {[tool.name for tool in self._tools]}")
 
@@ -118,9 +125,15 @@ class ToolManager:
         return None
 
     def get_forced_tools(self) -> list[ForcedToolOption]:
-        return [ForcedToolOption(t.name) for t in self._tools if t.name in self._tool_choices]
+        return [
+            ForcedToolOption(t.name)
+            for t in self._tools
+            if t.name in self._tool_choices
+        ]
 
-    def get_tool_definitions(self) -> list[LanguageModelToolDescription]:
+    def get_tool_definitions(
+        self,
+    ) -> list[LanguageModelTool | LanguageModelToolDescription]:
         return [tool.tool_description() for tool in self._tools]
 
     def get_tool_prompts(self) -> list[ToolPrompts]:
@@ -137,19 +150,19 @@ class ToolManager:
         )
         num_tool_calls = len(tool_calls)
 
-        if num_tool_calls > self._config .max_tool_calls:
+        if num_tool_calls > self._config.max_tool_calls:
             self._logger.warning(
                 (
                     "Number of tool calls %s exceeds the allowed maximum of %s."
                     "The tool calls will be reduced to the first %s."
                 ),
                 num_tool_calls,
-                self._config .max_tool_calls,
-                self._config .max_tool_calls,
+                self._config.max_tool_calls,
+                self._config.max_tool_calls,
             )
-            tool_calls = tool_calls[: self._config .max_tool_calls]
+            tool_calls = tool_calls[: self._config.max_tool_calls]
 
-        tool_call_responses = await self._execute_parallelized(tool_calls=tool_calls)
+        tool_call_responses = await self._execute_parallelized(tool_calls)
         return tool_call_responses
 
     async def _execute_parallelized(
@@ -194,6 +207,9 @@ class ToolManager:
             tool_response: ToolCallResponse = await tool_instance.run(
                 tool_call=tool_call
             )
+            evaluation_checks = tool_instance.evaluation_check_list()
+            self._tool_evaluation_check_list.update(evaluation_checks)
+
             return tool_response
 
         return ToolCallResponse(